requirements.txt
yfinance>=0.2.36
gspread>=6.0.0
google-auth>=2.28.0
requests>=2.31.0
pandas>=2.0.0

# Angel One — uncomment when using Angel One
# smartapi-python>=1.3.5
# pyotp>=2.9.0

# Zerodha — uncomment when using Zerodha
# kiteconnect>=5.0.1
